% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.xcms.dda.peaks.R
\name{assign.xcms.dda.peaks}
\alias{assign.xcms.dda.peaks}
\title{assign.xcms.dda.peaks}
\usage{
assign.xcms.dda.peaks(
  ramclustObj = NULL,
  xcms.feature.msmsObj = NULL,
  cmpd = NULL
)
}
\arguments{
\item{ramclustObj}{ramclustR object to annotate.}

\item{cmpd}{integer: vector defining compound numbers to annotated.  if NULL (default), all compounds}

\item{min.cor}{numeric: minimum dataset-wide correlational r-value to be considered for assignment.}
}
\value{
$M:  The inferred molecular weight of the compound giving rise to the each spectrum
}
\description{
assign.xcms.dda.peaks
}
\details{
This function is developed to assign fragment ions to ramclustR compounds after performing peak finding in the MSe (DIA MS/MS) data from centWave XCMS output.
}
\author{
Corey Broeckling
}
\concept{RAMClustR}
\concept{mass spectrometry}
\concept{metabolomics}
\concept{xcms}
