% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.lcms.library.R
\name{build.lcms.library}
\alias{build.lcms.library}
\title{lib.search}
\usage{
build.lcms.library(
  lib.directory = NULL,
  ri.file.name = NULL,
  column = "T3",
  lib.name = "cannabinoids-t3",
  trim.to.precursor = TRUE,
  intensity.filter = NULL
)
}
\arguments{
\item{lib.directory}{allowable product ion mass error in spectral matching in parts per million.  generally, use ~ 15 for TOF, 500 for quad.}

\item{ri.file.name}{allowable precursor ion mass error in spectral matching in parts per million.  generally, use ~ 15 for TOF, 500 for quad. Note though that quad isolation is NOT accurate mass, so you could consider using a larger window to account for potential chimeric spectra.}

\item{column}{retention time (or index) sigma for adjusting spectral match scores when retention similarity is used. use similarly to RAMClustR retention time similarity calculation.}

\item{lib.name}{experimental spectra, as an R object of class 'Spectra' produced from the Spectra package}

\item{trim.to.precursor}{logical. should MS/MS mass range be trimmed to below the precursor mass? If TRUE, values above the precursor mass + 4 m/z are discarded}

\item{get.pubchem}{logical. should pubchem parameters be retrieved for inclusion in spectrum output? If TRUE, INCHIKEY, Smiles, formula, and a few other paramaters are added to the msp format. while NIST cannot use all these, R Spectra can import them.}
}
\value{
nothing - exports an msp formatted file named lib.name.msp
}
\description{
spectral searching using Spectra::compare function(s)
}
\details{
function for building an LC-MS precursor (or not) MS/MS library in msp format for convervion to NIST format or searching within R.
}
\author{
Corey Broeckling
}
\concept{RAMClustR}
\concept{mass spectrometry}
\concept{metabolomics}
\concept{ramclustR}
\concept{spectral matching}
