% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.cmpd.get.classyfire.R
\name{classyfire.sunburst.plot}
\alias{classyfire.sunburst.plot}
\title{classyfire.sunburst.plot}
\usage{
classyfire.sunburst.plot(
  ramclustObj = RC,
  subset.cmpds = rep(TRUE, length(ramclustObj$ann)),
  out.file.name = NULL
)
}
\arguments{
\item{ramclustObj}{ramclustR object with classyfire data}

\item{subset.cmpds}{optionally, vector of integers of which compounds to include.}

\item{out.file.name}{output file name, optionally.}
}
\value{
nothing - exported files
}
\description{
interactive sunburst plot for exploring the classyfire heirarchy.
}
\details{
interactive sunburst plot for exploring the classyfire heirarchy.
}
\references{
Djoumbou Feunang Y, Eisner R, Knox C, Chepelev L, Hastings J, Owen G, Fahy E, Steinbeck C, Subramanian S, Bolton E, Greiner R, and Wishart DS. ClassyFire: Automated Chemical Classification With A Comprehensive, Computable Taxonomy. Journal of Cheminformatics, 2016, 8:61. DOI: 10.1186/s13321-016-0174-y
}
\author{
Corey Broeckling
}
\concept{MSFinder}
\concept{RAMClustR}
\concept{classyFire}
\concept{clustering}
\concept{feature}
\concept{mass spectrometry}
\concept{metabolomics}
\concept{ramclustR}
\concept{xcms}
