% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degolm.R
\name{degolm}
\alias{degolm}
\title{convenient editing of ramclustObj sample names}
\usage{
degolm(ramclustObj = RC)
}
\arguments{
\item{ramclustObj}{ramclustR object to edit}
}
\value{
returns an updated ramclustR object, with truncated GLM names, in addition to inchikey when available.
}
\description{
convenient editing of ramclustObj sample names
}
\details{
convenience function editing of sample names within R to correct errors or add additional factors. outputs a pdf plot and table in the 'spectra' directory which can be used to refine annotations.
}
\author{
Corey Broeckling
}
\concept{RAMClustR}
