% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editSampleNames.R
\name{editSampleNames}
\alias{editSampleNames}
\title{convenient editing of ramclustObj sample names}
\usage{
editSampleNames(ramclustObj = RC, delim = "-", cmpdlabel = "cmpd")
}
\arguments{
\item{ramclustObj}{ramclustR object to edit}

\item{delim}{character; "-" by default - the delimiter for parsing sample names to factors}

\item{cmpdlabel}{= "cmpd";  label the data with the annotation. can also be set to 'ann' for column names assigned as annotatins.}
}
\value{
returns a list of length 3: $design is the experimental sample factors after parsing by the delim, $data is the dataset, $full.data is merged $des and $data data.frames.
}
\description{
convenient editing of ramclustObj sample names
}
\details{
convenience function editing of sample names within R to correct errors or add additional factors
}
\author{
Corey Broeckling
}
\concept{RAMClustR}
