% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.cmpd.get.classyfire.R
\name{get.classyfire.local}
\alias{get.classyfire.local}
\title{get.classyfire.local}
\usage{
get.classyfire.local(
  ramclustObj = NULL,
  chemontid = NULL,
  chemont.obo.location = NULL
)
}
\arguments{
\item{ramclustObj}{ramclustR object to ClassyFy.  Must supply one of either ramclustObj or inchikey (see below)}

\item{chemontid}{vector of chem ontology IDs.}

\item{chemont.obo.location}{location of the chemontology OBO file.}
}
\value{
returns a ramclustR object.  new dataframe in $classyfire slot with rows equal to number of compounds.
}
\description{
use classyfire web API to look up full ClassyFire hirarchy for each inchikey
}
\details{
The get chem ontology heirarchy from chemontid. references local .obo file for faster lookup
}
\references{
Djoumbou Feunang Y, Eisner R, Knox C, Chepelev L, Hastings J, Owen G, Fahy E, Steinbeck C, Subramanian S, Bolton E, Greiner R, and Wishart DS. ClassyFire: Automated Chemical Classification With A Comprehensive, Computable Taxonomy. Journal of Cheminformatics, 2016, 8:61. DOI: 10.1186/s13321-016-0174-y
}
\author{
Corey Broeckling
}
\concept{MSFinder}
\concept{RAMClustR}
\concept{classyFire}
\concept{clustering}
\concept{feature}
\concept{mass spectrometry}
\concept{metabolomics}
\concept{ramclustR}
\concept{xcms}
