% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.pubchem.annotations.all.R
\name{get.pubchem.annotations.all}
\alias{get.pubchem.annotations.all}
\title{get.pubchem.annotations.all}
\usage{
get.pubchem.annotations.all(annotation = "")
}
\arguments{
\item{annotation}{character.  must be an exact text match (case sensitive) to values as reported here: https://pubchem.ncbi.nlm.nih.gov/rest/pug/annotations/headings/JSON}
}
\value{
returns a data frame with pubchem CID, compound name, data source, and returned value.  All data except CID are returned as character (text) as many values are a mix of text and numeric characters.
}
\description{
use PUG view to retrieve, for example, all pubchem values for 'Dissociation Constants', or other annotations.  all annotations found: https://pubchem.ncbi.nlm.nih.gov/rest/pug/annotations/headings/JSON
}
\details{
useful for moving from chemical name to digital structure representation. greek letters are assumed to be 'UTF-8' encoded, and are converted to latin text before searching.   if you are reading in your compound name list, do so with 'encoding' set to 'UTF-8'.
}
\author{
Corey Broeckling
}
