% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golm.inchikey.R
\name{golm.inchikey}
\alias{golm.inchikey}
\title{retrieve inchikeys GOLM msp}
\usage{
golm.inchikey(
  ramclustObj = RC,
  msp = "R:/RSTOR-PMF/Projects/Broeckling_Corey/GMD_20111121_VAR5_ALK_MSP.txt"
)
}
\arguments{
\item{ramclustObj}{ramclustR object to edit}
}
\value{
returns an updated ramclustR object with inchikeys when available.
}
\description{
retrieve inchikeys GOLM msp
}
\details{
convenience function for editing RC object.  matches compound name to those in msp file, returns compound (metabolite) inchikey when available.
}
\author{
Corey Broeckling
}
\concept{RAMClustR}
