% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-zip.R
\name{make.zip.files}
\alias{make.zip.files}
\title{zip selected directories and move to 'report' directory.}
\usage{
make.zip.files(
  directories = c("datasets", "spectra", "QC", "stats"),
  do.raw = TRUE,
  raw.file.extensions = c(".cdf", ".mzML", ".mzXML", ".cmbx", ".PRO", ".raw")
)
}
\arguments{
\item{raw.file.extensions}{- character vector. which file extensions represent raw files.  Default: c(".cdf", ".mzML", ".mzXML", ".PRO", ".cmbx", ".raw").}

\item{directories:}{character vector.  Which directories should be zipped. Only files in these directories will be chosen.}

\item{do.raw:}{logical - should a zip file of raw data files be generated?}
}
\value{
Nothing at all
}
\description{
zip selected directories and move to 'report' directory.
}
\details{
convenience function for zipping raw data and processed data (two separate zip files)
}
\author{
Corey Broeckling
}
\concept{RAMClustR}
