% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.functions.R
\name{normalize.intensity}
\alias{normalize.intensity}
\title{lib.functions}
\usage{
normalize.intensity(x, ...)
}
\arguments{
\item{mz.min}{numeric. lower m/z range limit for filter.by.mass function}

\item{mz.max}{numeric. upper m/z range limit for filter.by.mass function}

\item{min.int}{numeric. remove.low.intensity.signals function. value between 0 and 1 setting the threshold for the lower limit of intensity relative to the base peak intensity.}
}
\description{
lib.functions
}
\author{
Corey Broeckling
}
\concept{spectral matching}
