% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc.bio.to.fiora.R
\name{pc.bio.to.fiora}
\alias{pc.bio.to.fiora}
\title{pc.bio.to.fiora}
\usage{
pc.bio.to.fiora(
  cid = vector(length = 0, mode = "integer"),
  pc.data.frame = NULL,
  pc.directory = "R:/RSTOR-PMF/Software/db/met.db/20241216/",
  use.parent.cid = FALSE,
  precursor.type = c("[M+H]+"),
  ce = c(10, 20, 40),
  instrument.type = "HCD",
  remove.inorganics = TRUE,
  remove.salts = TRUE,
  remove.redundant = TRUE,
  remove.invalid.smiles = FALSE,
  mass.range = c(50, 2000),
  subset.cmpds = 1:nrow(out),
  batch.output.size = 5000,
  output.filename = paste0(pc.directory, "fiora.input.csv")
)
}
\arguments{
\item{cid}{vector of integers.  should be subset of all CIDs found in the $cid slot of the pc.data.frame.}

\item{pc.directory}{directory from which to load pubchem .Rdata files}

\item{use.parent.cid}{logical. should CIDs be replaced with parent CIDs?}

\item{remove.inorganics}{logical.  if TRUE, removes all molecules which have exactly one element, or have two or more and have zero carbon or oxygen. In this way, phosphorus is removed, but phosphate stays.}

\item{bio.sources}{vector of source names from which to extract pubchem CIDs.  all can be found here: https://pubchem.ncbi.nlm.nih.gov/sources/.  deafults to c("Metabolomics Workbench", "Human Metabolome Database (HMDB)", "ChEBI", "LIPID MAPS",  "MassBank of North America (MoNA)")}

\item{use.pathways}{logical.  should all CIDs from any biological pathway data be used?}

\item{use.taxid}{logical.  should all CIDs associated with a taxonomic identifier (taxid) be used?}

\item{get.properties}{logical. if TRUE, will return pubchem properties.  Either a short list of readily available properties including by default XLogP, TPSA, HBondDonorCount and HBondAcceptorCount, or a comprehensive list, if all.props = TRUE.}

\item{threads}{integer. how many threads to use when calculating rcdk properties.  parallel processing via DoParallel and foreach packages.}
}
\value{
a data frame containing pubchem CID, title, formula, monoisotopic molecular weight, inchikey, smiles, cas, optionally pubchem properties
}
\description{
utilizes downloaded and properly formatted local pubchem data created by 'get.pubchem.ftp' function
}
\details{
utilizes downloaded and properly formatted local pubchem data created by 'get.pubchem.ftp' function
}
\author{
Corey Broeckling
}
