% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmf.designExperiment.R
\name{pmf.defineExperiment}
\alias{pmf.defineExperiment}
\title{pmf.defineExperiment}
\usage{
pmf.defineExperiment()
}
\value{
returns an R object suitable for use as an ExpDes input to ramclustR() function.
}
\description{
set up exeperimental design and instrument description
}
\details{
differs from RAMClustR define experiment in having preconfigured instruments and slots for factor descriptiosn.
}
\concept{RAMClustR}
\concept{ramclustR}
