% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmf.directory.report.R
\name{pmf.directory.report}
\alias{pmf.directory.report}
\title{returns a dataframe of all directories and their size in megabytes and last accessed (in days).}
\usage{
pmf.directory.report(dir = getwd())
}
\arguments{
\item{dir}{directory to summarize.  default is getwd().  can also be set - i.e. "C:/Users/username/Desktop/temp"}
}
\value{
returns an dataframe with three columns, including directory path, directory size in megabytes and the minimum number of days since any file within that directory was accessed.  'accessed' includes any new files being written or any old file being modified and saved.
}
\description{
returns a dataframe of all directories and their size in megabytes and last accessed (in days).
}
\details{
convenience function determining which directories might be archived.
}
\author{
Corey Broeckling
}
