% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmf-anova.R
\name{pmfanova}
\alias{pmfanova}
\title{pmfanova}
\usage{
pmfanova(ramclustObj = RC, anova.name = NULL,
  which.data = "SpecAbund", subset = c(""), anova.call = NULL,
  posthoc = NULL, effectsplots = TRUE, delim = "-", plots = TRUE,
  pcut = 0.05, p.adj = "BH", which.quan = NULL)
}
\arguments{
\item{ramclustObj}{ramclustR object to perform ANOVA on}

\item{anova.name}{character; optionally assign a name to the anova. valuable if you wish to perform more than one ANOVA on a sample set. If NULL, name will be autogenerated from the anova.call and subset information.}

\item{which.data}{character; which dataset (SpecAbund or SpecAbundAve) to perform PCA on.}

\item{subset}{character vector; must be even length.  If you wish to perform ANOVA only when your factor called 'treatment' is a 'trt' sample and when 'time' is '3', then you would use is i.e. c("treatment", "trt", "time", "3"). vector length must always be even, and with 'factor' followed by 'level'.}

\item{scale}{character; default = 'pareto'.  will also accept 'uv' or 'none'}

\item{which.factors}{character vector; i.e. which.factors = c("treatment", "time").   which factors should be used for coloring PCA plots?}

\item{num.factors}{which factors should be treated as numeric? must be subset of 'which.factors'. i.e. c("time")}

\item{label.by}{how should metabolites columns be labelled? one of 'ann' or 'cmpd', typically.}

\item{npc}{"auto" by default (recommended).  This will autoselect number of PCs to use.  Can also be set to any integer value to force more PCs.}
}
\value{
returns a ramclustR object.  new R object in $pca slot. Optionally, new R object in $AuerGervini slot if npc = "auto".
}
\description{
perform principle compponent analysis on ramclustR object dataset, export plots
}
\details{
This function performs standard or mixed model anova using a ramclusR object as input
}
\author{
Corey Broeckling
}
\concept{RAMClustR}
