% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmf-xcms_new.R
\name{pmfxcms.2}
\alias{pmfxcms.2}
\title{pmfxcms}
\usage{
pmfxcms.2(
  filetype.ms1 = "_01.mzML",
  filetype.ms2 = "_02.mzML",
  sequence = "seq.csv",
  delim = "-",
  ms.res = 30000,
  peak.width = c(3, 30),
  sn = 5,
  minfrac = 0.4,
  bw.pre = 3,
  bw.post = 1.2,
  n.cores = 4
)
}
\arguments{
\item{filetype.ms1}{File extension of raw data for MS level data.  This extention will be scanned in all raw data files in working directory and will be added to filenames in sequence .csv file}

\item{filetype.ms2}{File extension of raw data for MSe level data.  If not MSe data, set to NA.  This extention will be scanned in all raw data files in working directory and will be added to filenames in sequence .csv file}

\item{sequence}{.csv file containing filenames (no extension) in first column, and sample name and/or factor names in remaining columns.}

\item{delim}{symbol used to delimit factors in sequence file.  "-" by default.}

\item{ms.res}{numerical value representing mass resolution of the MS system.  if NA, assumed to be quad data (low res). if numeric, used to set mass tolerances for xcms.}

\item{peak.width}{numerical vector of length two giving peak width range in seconds i.e. c(3, 30)}

\item{sn}{signal to noise threshold for peak detection}

\item{minfrac}{minimum proportion of samples in which a feature must be present (at one MS level). i.e. 0.4 is 4 out of 10.}

\item{bw.pre}{bandwidth for pre-rt correction grouping.  suggested value is 3, but if there is a good deal of retention drift over the sample set a larger value should be used.}

\item{bw.post}{bandwidth for post-rt correction grouping.  suggested values is 1.5.}
}
\value{
returns an xcms object
}
\description{
A wrapper function for XCMS using common PMF platforms.
}
\details{
This function uses a template file as input to guide XCMS parameter selection.

Peak detection using centwave (LC-TOF) or matchedFilter (GC-Quad), density based peak grouping, loess rt correction, regrouping, and fillPeaks.

Additionally performs unsupervised removal of ultrawide chromatgoraphic peaks and outlier detection and removal after peak detection step.

an XCMS object is returned.
}
\author{
Corey Broeckling
}
\concept{RAMClustR xcms}
