% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchem.functions.R
\name{pubchem.taxon.metabolome}
\alias{pubchem.taxon.metabolome}
\title{get.taxon.cids}
\usage{
pubchem.taxon.metabolome(
  taxid = NULL,
  taxstring = NULL,
  sub.taxa.n = 100,
  ncbi.api.key = NULL
)
}
\arguments{
\item{taxid}{integer NCBI taxid for the taxon to search.  Must supply one of taxid OR taxstring - not both. can be a vector of integers.}

\item{taxstring}{character.  Taxonomy string to search.  tested with Genus species and or Genus alone.   Must supply one of taxid OR taxstring - not both. can be a vector of character strings (genus names).}

\item{sub.taxa.n}{integer value for the number of subtaxa to consider.  Note that if the sub.taxa.n value is less the the availabe number of subtaxa, only the first sub.taxa.n values, as reported by rentrez, are returned.  If you require specific subtaxa, you should call those taxids explicitly to ensure those results are returned.}

\item{ncbi.api.key}{character. optional NCBI API key to enable 10 calls per second (3 by default).  Note that this function is not parallelized, but if you wish to parallelize you may require use of a key to take full advantage.}
}
\value{
returns a vector of integer pubchem cids.
}
\description{
use pubchem rest to retreive pubchem CIDS known to be found in a given species.  NCBI taxid should be used as input.  i.e. Homo sapiens subsp. 'Denisova' is taxid 741158
}
\details{
this function enables return of a list of pubchem CIDs which can be used for prioritizing annotations.  If a genus level taxid is selected, setting the sub.taxa.n option > 0 will return metabolites associated with that taxid and all (assuming n is large enough) subtaxa.  i.e. seting taxid to 9605 (genus = 'Homo') will return metabolites associated with Homo sapiens, Homo heidelbergensis, Homo sapiens subsp. 'Denisova', etc.
}
\author{
Corey Broeckling
}
