% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.get.xcms.msms.R
\name{rc.get.xcms.msms}
\alias{rc.get.xcms.msms}
\title{rc.get.xcms.msms}
\usage{
rc.get.xcms.msms(
  xcmsObj = x.data,
  ramclustObj = RC,
  expandRt = 0,
  expandMz = 0.3,
  ppm = 15,
  centroid = TRUE,
  in.memory = TRUE,
  aggregate.spectra = TRUE,
  weighted = TRUE,
  BPPARAM = SnowParam(2)
)
}
\arguments{
\item{xcmsObj}{xcms Object.  Should be same object that was used to generate RAMCLustR object and contain MS2 spectra.}

\item{expandMz}{logical.  mz expansion for matching precusor ion.  necessary since ion isolation for MS/MS is generally performed using a quad or ion trap, and isolation width is much larger than mass measurement error of TOF/Orbitrap. default = 0.3.}

\item{ppm}{numerical.  ppm similarity required to combine two signals.}

\item{centroid}{logical. should CIDs be replaced with parent CIDs?}

\item{in.memory}{logical. if TRUE, spectra will be stored in memory.  if false, stored on disk.  in.memory will be faster, assuming sufficient memory is available.}

\item{aggregate.spectra}{logical. should multiple spectra for the same feature be aggregated into a single spectrum?  Default = TRUE.}

\item{weighted}{logical.  should weighting be used during spectrum aggregation if aggregate.spectra is TRUE?  Weights are signal intensity.}

\item{BPPARAM}{BiocParallel object, passed to "Spectra" and "XCMS" functions. default = SnowParam(2)}

\item{ramClustObj}{RAMClustR object.}

\item{expantRt}{logical.  rt expansion for matching precusor ion. default = 0.}
}
\value{
RAMClustR object.
}
\description{
utilizes downloaded and properly formatted local pubchem data created by 'get.pubchem.ftp' function
}
\details{
takes as input a ramclustR object and an XCMS object.  Extracts MS/MS data from xcms object to support annotation of RAMClustR feature groups.
}
\author{
Corey Broeckling
}
