% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.run.msfinder.R
\name{rc.run.msfinder}
\alias{rc.run.msfinder}
\title{rc.run.msfinder}
\usage{
rc.run.msfinder(
  msfinder.dir = "C:/MSFinder/msfinder",
  input.dir = "R:/RSTOR-PMF/Temp/mat",
  param.file = "R:/RSTOR-PMF/Software/MSFinder/MSFINDER.INI"
)
}
\arguments{
\item{msfinder.dir}{directory containing the msfinder executable file}

\item{input.dir}{directory containing all .mat files}

\item{param.file}{file (with full path) to msfinder paramater file}
}
\value{
nothing - output written to drive.
}
\description{
run MSFinder for LC-MS/MS annotation
}
\details{
function to call msfinder on linux server
}
\author{
Corey Broeckling
}
\concept{MSFinder}
