% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportSampleNames.R
\name{reportSampleNames}
\alias{reportSampleNames}
\title{convenient reporting of sample name and design}
\usage{
reportSampleNames(
  ramclustObj = RC,
  which.data = "SpecAbund",
  delim = "-",
  cmpdlabel = "cmpd",
  filename = NULL
)
}
\arguments{
\item{ramclustObj}{ramclustR object summarize}

\item{which.data}{character; which dataset (SpecAbund or SpecAbundAve) to perform PCA on.}

\item{delim}{character; "-" by default - the delimiter for parsing sample names to factors}

\item{cmpdlabel}{= "cmpd";  label the data with the annotation. can also be set to 'ann' for column names assigned as annotatins.}
}
\value{
returns a ramclustR Object, with sample summary appended to 'sampleSummary'slot.
}
\description{
convenient reporting of sample name and design
}
\details{
convenience function reporting of sample names.  writes text file with sample summary.
}
\author{
Corey Broeckling
}
\concept{RAMClustR}
