% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax.filter.pubchem.bio.R
\name{tax.filter.pubchem.bio}
\alias{tax.filter.pubchem.bio}
\title{tax.filter.pubchem.bio}
\usage{
tax.filter.pubchem.bio(
  taxid = NULL,
  pubchem.bio.db = NULL,
  pc.directory = "R:/RSTOR-PMF/Software/db/met.db/20241216"
)
}
\arguments{
\item{taxid}{integer vector.  containing taxid values (from NCBI taxonomy database).  can be vector of any length.}

\item{pubchem.bio.db}{data.table, generally created by the build.pubchem.bio function, or data.table with one column named 'cid' which has pubchem cid values in integer format.}

\item{pc.directory}{directory from which to load pubchem .Rdata files}

\item{tax.sources}{vector. which taxonomy sources should be used?  defaults to c("LOTUS - the natural products occurrence database", "The Natural Products Atlas", "KNApSAcK Species-Metabolite Database", "Natural Product Activity and Species Source (NPASS)").}
}
\value{
a data frame containing pubchem CID, title, formula, monoisotopic molecular weight, inchikey, smiles, cas, optionally pubchem properties
}
\description{
utilizes downloaded and properly formatted local pubchem data created by 'get.pubchem.ftp' function to filter a dataset created by 'build.pubchem.bio' function
}
\details{
utilizes downloaded and properly formatted local pubchem data created by 'get.pubchem.ftp' function
}
\author{
Corey Broeckling
}
