% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-file-summary.R
\name{write.file.summary}
\alias{write.file.summary}
\title{output file descriptions for reporting}
\usage{
write.file.summary()
}
\value{
NA
}
\description{
output file descriptions for reporting
}
\details{
convenience function providing a brief description of the common files delivered with the report. Looks in the working directory for the standard directory and file names and writes a new file to the report directory called 'file.descriptions.txt'.
}
\author{
Corey Broeckling
}
\concept{RAMClustR}
