% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmf-anova.R
\name{pmfanova}
\alias{pmfanova}
\title{pmfanova}
\usage{
pmfanova(
  ramclustObj = RC,
  anova.name = NULL,
  which.data = "SpecAbund",
  label.by = "cmpd",
  subset = c(""),
  anova.call = NULL,
  posthoc = NULL,
  effectsplots = TRUE,
  delim = "-",
  plots = TRUE,
  pcut = 0.05,
  p.adj = "BH",
  which.quan = NULL,
  filter = TRUE,
  summary.statistics = TRUE,
  output.summary = TRUE,
  remove.na.factors = FALSE
)
}
\arguments{
\item{ramclustObj}{ramclustR object to perform ANOVA on}

\item{anova.name}{character; optionally assign a name to the anova. valuable if you wish to perform more than one ANOVA on a sample set. If NULL, name will be autogenerated from the anova.call and subset information.}

\item{which.data}{character; which dataset (SpecAbund or SpecAbundAve) to perform PCA on.}

\item{label.by}{how should metabolites columns be labelled? one of 'ann' or 'cmpd', typically.}

\item{subset}{character or integer vector; if character, must be even length.  If you wish to perform ANOVA only when your factor called 'treatment' is a 'trt' sample and when 'time' is '3', then you would use  i.e. c("treatment", "trt", "time", "3"). vector length must always be even, and with 'factor' followed by 'level'.  If an integer vector is provided, only row numbers matching those integers are retained.}

\item{anova.call}{character;  the model you wish to run.  i.e. "Trt * Time" or "Trt * Time + (1|Block)"}

\item{posthoc}{character; if you wish to perform post hoc testing, define the posthoc test to use.  i.e. if anova.call is "Trt * Time" you can use "Trt|Time" to specific 'treatment within time' comparisons.}

\item{effectsplots}{logical; if TRUE, plot effect plots in pdf format.  If plotting fails, can set to 'FALSE' to get tabular results only.}

\item{delim}{character; what is the character delimiting factors in the sample names.  generally '-'}

\item{plots}{logical; if plotting is any issue set this to FALSE.}

\item{pcut}{numeric; what is the pvalue cutoff for determining significance?  Used only for plotting.}

\item{p.adj}{character; what p.adjust method should be used.  generally 'fdr' or 'bh' (equivalent). see ?p.adjust}

\item{filter}{logical, TRUE by default. when $cmpd.use slot is present (from rc.cmpd.filter.cv function), only cmpds that passed cv filtering are used. If you wish to change that behavior, rerun the rc.cmpd.filter.cv function with a really high CV threshold.}

\item{summary.statistics}{logical, TRUE by default.  If true, calculate mean and standard deviation for each level as defined by the anova.call model.}

\item{output.summary}{logical, TRUE by default.  If true, saves full model results for all ANOVA as .Rdata file, and exports all summaries to a .txt file.}

\item{which.quan;}{chracter vector.  which factors should be interpreted as quantitative?  i.e. c("time", "dose")}
}
\value{
returns a ramclustR object.  new R object in $pca slot. Optionally, new R object in $AuerGervini slot if npc = "auto".
}
\description{
perform principle compponent analysis on ramclustR object dataset, export plots
}
\details{
This function performs standard or mixed model anova using a ramclusR object as input
}
\author{
Corey Broeckling
}
\concept{RAMClustR}
