% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmf-pca.R
\name{pmfpca}
\alias{pmfpca}
\title{pmfpca}
\usage{
pmfpca(
  ramclustObj = RC,
  which.data = "SpecAbund",
  scale = "pareto",
  subset = c(""),
  which.factors = NULL,
  num.factors = NULL,
  label.by = "cmpd",
  npc = "auto",
  ag.summary.plot = FALSE,
  filter = FALSE
)
}
\arguments{
\item{ramclustObj}{ramclustR object to perform PCA on}

\item{which.data}{character; which dataset (SpecAbund or SpecAbundAve) to perform PCA on.}

\item{scale}{character; default = 'pareto'.  will also accept 'uv' or 'none'}

\item{subset}{character or integer vector; if character, must be even length.  If you wish to perform PCA only when your factor called 'treatment' is a 'trt' sample and when 'time' is '3', then you would use  i.e. c("treatment", "trt", "time", "3"). vector length must always be even, and with 'factor' followed by 'level'.  If an integer vector is provided, only row numbers matching those integers are retained.}

\item{which.factors}{character vector; i.e. which.factors = c("treatment", "time").   which factors should be used for coloring PCA plots?}

\item{num.factors}{which factors should be treated as numeric? must be subset of 'which.factors'. i.e. c("time")}

\item{label.by}{how should metabolites columns be labelled? one of 'ann' or 'cmpd', typically.}

\item{npc}{"auto" by default (recommended).  This will autoselect number of PCs to use.  Can also be set to any integer value to force more PCs.}

\item{filter}{logical, TRUE by default. when $cmpd.use slot is present (from rc.cmpd.filter.cv function), only cmpds that passed cv filtering are used. If you wish to change that behavior, rerun the rc.cmpd.filter.cv function with a really high CV threshold.}
}
\value{
returns a ramclustR object.  new R object in $pca slot. Optionally, new R object in $AuerGervini slot if npc = "auto".
}
\description{
perform principle compponent analysis on ramclustR object dataset, export plots
}
\details{
This function uses the native prcomp() function in R to perform PCA analysis.

Automatic selection of the number of principle components using the AuerGervini method is enabled by the PCDimension and ClassDiscovery packages.

R PCA AuerGervini objects are attached to the ramclustR object, and summary plots and csv files are written to the stats/pca directory in the working directory

a methods narrative is also appended to the $history slot
}
\author{
Corey Broeckling
}
\concept{RAMClustR}
