% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startProject.R
\name{startProject}
\alias{startProject}
\title{startProject}
\usage{
startProject(
  prep.batch.size = 48,
  run.batch.size = 96,
  QC = 6,
  randomize = TRUE,
  stack = FALSE,
  prep.blanks = 3,
  LTR = 0,
  remove.invariant.factors = TRUE,
  destination.dir = "R:/RSTOR-PMF/Projects/",
  sub.project.name = NULL
)
}
\arguments{
\item{prep.batch.size}{integer - what is the sample preparation batch size?  default = 48.}

\item{run.batch.size}{integer - what is the analytical batch size? default = 96.}

\item{QC}{integer - every 'QC' injections will be a QC sample.  Default = 6}

\item{randomize}{logical - should prep and run order be randomized?}

\item{stack}{should the injection sample list be set up for stacked injections (TOF platforms only, currenly)}

\item{prep.blanks}{integer - how many extraction blanks per prep batch? prep.blanks will be inserted into the prep sequence in random sequence position(s).}

\item{LTR}{logical - If TRUE, an long term reference sample line will be included at the beginning and end of each run batch, with each LTR followed by an inserted solvent blank.}

\item{destination.dir}{valid path to directory target output directory, by default: "R:/RSTOR-PMF/Projects/"}
}
\value{
returns nothing, files written to directory selected by user.
}
\description{
startProject
}
\details{
take info from iLab and iLab sample submission form and create sample prep and data acquisition templates and directories on the R:
}
\author{
Corey Broeckling
}
