% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startProject.R
\name{startProject}
\alias{startProject}
\title{startProject}
\usage{
startProject(prep.batch.size = 48, run.batch.size = 96, QC = 6,
  randomize = TRUE, stack = FALSE, extraction.blanks = 3,
  solvent.blanks = 3)
}
\arguments{
\item{prep.batch.size}{integer - what is the sample preparation batch size?  default = 48.}

\item{run.batch.size}{integer - what is the analytical batch size? default = 96.}

\item{QC}{integer - every 'QC' injections will be a QC sample.  Default = 6}

\item{stack}{should the injection sample list be set up for stacked injections (TOF platforms only, currenly)}

\item{extraction.blanks}{integer - currently not used. (how many extraction blanks per prep batch?)}

\item{solvent.blanks}{integer - currently not used.  (how many solvent blanks per run batch? )}
}
\value{
returns a list of length 3: $design is the experimental sample factors after parsing by the delim, $data is the dataset, $full.data is merged $des and $data data.frames.
}
\description{
startProject
}
\details{
take info from iLab and iLab sample submission form and create sample prep and data acquisition templates and directories on the R:
}
\author{
Corey Broeckling
}
\concept{RAMClustR}
