% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startProject.R
\name{startProject}
\alias{startProject}
\title{startProject}
\usage{
startProject(
  prep.batch.size = 48,
  run.batch.size = 96,
  QC = 6,
  randomize = TRUE,
  stack = FALSE,
  prep.blanks = 3,
  LTR = 1,
  solvent.blanks = 0,
  destination.dir = "R:/RSTOR-PMF/Projects/",
  sub.project.name = NULL
)
}
\arguments{
\item{prep.batch.size}{integer - what is the sample preparation batch size?  default = 48.}

\item{run.batch.size}{integer - what is the analytical batch size? default = 96.}

\item{QC}{integer - every 'QC' injections will be a QC sample.  Default = 6}

\item{stack}{should the injection sample list be set up for stacked injections (TOF platforms only, currenly)}

\item{prep.blanks}{integer - how many extraction blanks per prep batch? prep.blanks will be inserted into the prep sequence in random sequence position(s).}

\item{LTR}{integer - how many long term reference material (LTR) vials per batch? default = 0.  LTR samples will be inserted at the start of each batch, and a solvent blank will ALwAYs be inserted after the LTR.}

\item{solvent.blanks}{integer - how many ADDITIONAL solvent blank vials  per batch? default = 0. solvent blanks will include one for each LTR automatically. Any additional solvent blanks specified here will be inserted randomly in each batch.}

\item{destination.dir}{valid path to directory target output directory, by default: "R:/RSTOR-PMF/Projects/"}
}
\value{
returns nothing, files written to directory selected by user.
}
\description{
startProject
}
\details{
take info from iLab and iLab sample submission form and create sample prep and data acquisition templates and directories on the R:
}
\author{
Corey Broeckling
}
